+++ home-work/src/repositories/posts-db-repository.ts
@@ -1,6 +1,7 @@
 import { blogsRepository } from './blogs-repository';
 import { clientDB } from './db';
 import { ObjectId } from 'mongodb';
+import { verifyId } from '../utils/utils';
 
 type PostType = {
   id?: string,
@@ -34,17 +35,19 @@ export const postsRepository = {
     });
   },
   async findPostById(id: string): Promise<PostType | null> {
-    const post: PostType | null  = await clientDB.collection<PostType>('posts').findOne({  _id: new ObjectId(id) });
-    if (post) {
-      return {
-        // @ts-ignore
-        id: post._id,
-        title: post.title,
-        shortDescription: post.shortDescription,
-        content: post.content,
-        blogId: post.blogId,
-        blogName: post.blogName
-      };
+    if (verifyId(id)) {
+      const post: PostType | null = await clientDB.collection<PostType>('posts').findOne({ _id: new ObjectId(id) });
+      if (post) {
+        return {
+          // @ts-ignore
+          id: post._id,
+          title: post.title,
+          shortDescription: post.shortDescription,
+          content: post.content,
+          blogId: post.blogId,
+          blogName: post.blogName
+        };
+      }
     }
     return null;
   },
@@ -70,13 +73,22 @@ export const postsRepository = {
   },
   async updatePost(id: string, post: {
     title: string, shortDescription: string, content: string
-  }): Promise<Array<PostType> | undefined> {
-    const updatedPost: any = await clientDB.collection<PostType>('posts').updateOne({  _id: new ObjectId(id)}, {"$set": {...post}})
-    return updatedPost;
+  }): Promise<Array<PostType> | undefined | null> {
+    if (verifyId(id)) {
+      const findPost = await postsRepository.findPostById(id);
+      if (findPost) {
+        const updatedPost: any = await clientDB.collection<PostType>('posts').updateOne({ _id: new ObjectId(id) }, { "$set": { ...post } })
+        return updatedPost;
+      }
+    }
+    return null;
   },
   async deletePost(id: string) : Promise<boolean> {
-    const deletedPost: any = await clientDB.collection<PostType>('posts').deleteOne({  _id: new ObjectId(id) });
-    return deletedPost?.deletedCount === 1;
+    if (verifyId(id)) {
+      const deletedPost: any = await clientDB.collection<PostType>('posts').deleteOne({ _id: new ObjectId(id) });
+      return deletedPost?.deletedCount === 1;
+    }
+    return false;
   },
   async deleteAllPosts(): Promise<boolean> {
     const deletedPosts: any = await clientDB.collection<PostType>('posts').deleteMany({});
